Move data entry functionality to new module
