+++ backend/src/lib.rs
@@ -1,6 +1,6 @@
 use std::error::Error;
 
-use crate::polling_station::DataError;
+use crate::data_entry::DataError;
 use axum::extract::rejection::JsonRejection;
 use axum::extract::FromRef;
 use axum::http::StatusCode;
@@ -16,6 +16,7 @@ use utoipa::ToSchema;
 #[cfg(feature = "openapi")]
 use utoipa_swagger_ui::SwaggerUi;
 
+pub mod data_entry;
 pub mod election;
 #[cfg(feature = "dev-database")]
 pub mod fixtures;
@@ -30,22 +31,22 @@ pub struct AppState {
 
 /// Axum router for the application
 pub fn router(pool: SqlitePool) -> Result<Router, Box<dyn Error>> {
-    let polling_station_routes = Router::new()
+    let data_entry_routes = Router::new()
         .route(
             "/:polling_station_id/data_entries/:entry_number",
-            post(polling_station::polling_station_data_entry_save),
+            post(data_entry::polling_station_data_entry_save),
         )
         .route(
             "/:polling_station_id/data_entries/:entry_number",
-            get(polling_station::polling_station_data_entry_get),
+            get(data_entry::polling_station_data_entry_get),
         )
         .route(
             "/:polling_station_id/data_entries/:entry_number",
-            delete(polling_station::polling_station_data_entry_delete),
+            delete(data_entry::polling_station_data_entry_delete),
         )
         .route(
             "/:polling_station_id/data_entries/:entry_number/finalise",
-            post(polling_station::polling_station_data_entry_finalise),
+            post(data_entry::polling_station_data_entry_finalise),
         );
 
     let election_routes = Router::new()
@@ -59,7 +60,7 @@ pub fn router(pool: SqlitePool) -> Result<Router, Box<dyn Error>> {
 
     let app = Router::new()
         .nest("/api/elections", election_routes)
-        .nest("/api/polling_stations", polling_station_routes);
+        .nest("/api/polling_stations", data_entry_routes);
 
     // Add a route to reset the database if the dev-database feature is enabled
     #[cfg(feature = "dev-database")]
@@ -90,17 +91,24 @@ pub fn create_openapi() -> utoipa::openapi::OpenApi {
             election::election_details,
             election::election_status,
             election::election_download_results,
-            polling_station::polling_station_data_entry_save,
-            polling_station::polling_station_data_entry_get,
-            polling_station::polling_station_data_entry_delete,
-            polling_station::polling_station_data_entry_finalise,
+            data_entry::polling_station_data_entry_save,
+            data_entry::polling_station_data_entry_get,
+            data_entry::polling_station_data_entry_delete,
+            data_entry::polling_station_data_entry_finalise,
         ),
         components(
             schemas(
                 ErrorResponse,
-                validation::ValidationResult,
-                validation::ValidationResultCode,
-                validation::ValidationResults,
+                data_entry::CandidateVotes,
+                data_entry::SaveDataEntryRequest,
+                data_entry::SaveDataEntryResponse,
+                data_entry::GetDataEntryResponse,
+                data_entry::DifferencesCounts,
+                data_entry::PoliticalGroupVotes,
+                data_entry::PollingStationResults,
+                data_entry::VotersCounts,
+                data_entry::VotersRecounts,
+                data_entry::VotesCounts,
                 election::Election,
                 election::ElectionCategory,
                 election::PoliticalGroup,
@@ -109,20 +117,13 @@ pub fn create_openapi() -> utoipa::openapi::OpenApi {
                 election::ElectionListResponse,
                 election::ElectionDetailsResponse,
                 election::ElectionStatusResponse,
-                polling_station::CandidateVotes,
-                polling_station::SaveDataEntryRequest,
-                polling_station::SaveDataEntryResponse,
-                polling_station::GetDataEntryResponse,
-                polling_station::DifferencesCounts,
-                polling_station::PoliticalGroupVotes,
-                polling_station::PollingStationResults,
                 polling_station::PollingStationType,
                 polling_station::PollingStationStatusEntry,
                 polling_station::PollingStationStatus,
                 polling_station::PollingStation,
-                polling_station::VotersCounts,
-                polling_station::VotersRecounts,
-                polling_station::VotesCounts,
+                validation::ValidationResult,
+                validation::ValidationResultCode,
+                validation::ValidationResults,
             ),
         ),
         tags(
